made a new system for login
Login now goes through requests.
A class implements the Login interface and implements it's two methods, loginSucces and loginFailure. then whenever the class wants to login, it calls the addLoginRequest(Login source) method with itself being the source. Then the login request is queed and if there are no others first, a new login screen will be created. Then when the login succeeds, it sends a reply to the source class